chore: unify executors
